Split code into multiple files Fixed some typescript validation Cleaned up imports
